+++ src/init.c
@@ -53,8 +53,7 @@ extern SEXP nr_text_basic_  (SEXP nr_, SEXP x_ , SEXP y_ , SEXP str_            
 extern SEXP nr_rect_    (SEXP nr_, SEXP x_ , SEXP y_ , SEXP w_, SEXP h_, SEXP fill_, SEXP color_, SEXP hjust_, SEXP vjust_);
 extern SEXP nr_circle_  (SEXP nr_, SEXP x_ , SEXP y_ , SEXP r_         , SEXP fill_, SEXP color_);
 extern SEXP nr_polyline_(SEXP nr_, SEXP x_ , SEXP y_                               , SEXP color_, SEXP close_);
-extern SEXP nr_polygon_ (SEXP nr_, SEXP x_ , SEXP y_                   , SEXP fill_, SEXP color_);
-extern SEXP nr_polygons_(SEXP nr_, SEXP x_ , SEXP y_, SEXP id_         , SEXP fill_, SEXP color_);
+extern SEXP nr_polygons_multi_(SEXP nr_, SEXP x_ , SEXP y_, SEXP id_         , SEXP fill_, SEXP color_);
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Color
@@ -94,8 +93,7 @@ static const R_CallMethodDef CEntries[] = {
   {"nr_rect_"      , (DL_FUNC) &nr_rect_      , 9},
   {"nr_circle_"    , (DL_FUNC) &nr_circle_    , 6},
   {"nr_polyline_"  , (DL_FUNC) &nr_polyline_  , 5},
-  {"nr_polygon_"   , (DL_FUNC) &nr_polygon_   , 5},
-  {"nr_polygons_"  , (DL_FUNC) &nr_polygons_  , 6},
+  {"nr_polygons_multi_"  , (DL_FUNC) &nr_polygons_multi_  , 6},
   
   {"colors_to_packed_cols_"   , (DL_FUNC) &colors_to_packed_cols_   , 1},
   {"packed_cols_to_hexcolors_", (DL_FUNC) &packed_cols_to_hexcolors_, 1},